back/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
mistralai==1.0.1
python-dotenv==1.0.0
pydantic==2.5.0
certifi>=2023.7.22
httpx>=0.25.0
requests>=2.31.0 